ubuntu-panel: Ensure dock position is respected in RTL
Dash-to-dock uses absolute LEFT/RIGHT position setting values even in RTL
mode, but in such case it adapts the dock position correctly.

So, when we're in RTL mode, a right-positioned dock is actually in the
left of the screen and vice-versa.

As per this, use a binding with custom mapping, to invert such values when
using an RTL language.

LP: #1876937