ubuntu-panel: Add proper matching between the dock monitor ID and mutter
The dock uses its own monitors numbering for settings to make it also work
with the gtk configurations that it ships.

As per this, the matching with the CcDisplayPanel numbers (that reflect
mutter IDs) can be wrong when the primary monitor is not the monitor with
index 0.

To fix this, include into the monitor placement list store another column
that contains the monitor index as the dock would compute it, and use this
value to set the settings and select the relative list item.

LP: #1877108