display: Improve fractional scaling toggling
When fractional scaling is enabled we can just apply it promptly, as it will
preserve the current integer settings anyways, in case it's disabled,
instead we may have to reset the monitor scaling to integer values and
require user confirmation.

To do this, move the settings handling inside the current configuration and
in case the fractional scaling settings is reverted, ensure that we show
the user the "Apply" button and reset the parameters before doing so.

While the fractional scaling disable is pending, we won't show any
fractional value, to the user.

LP: #1876891
LP: #1876894