display: Use renderer and layout to figure out fractional scaling state
No need to read gsettings to fetch the actual fractional scaling state, as
mutter gives us such informations on the real state, without having us to
deal with the settings state manually (also because the state may depend
on other aspects as well).

We can also allows to remove the GDK usage in the model and work correctly
in both X11 and wayland.