ubuntu: Don't try to fetch monitors if the config manager is not ready yet
In some cases the settings changes may be emitted when the display manager
configuration has not been initialized yet (being created asynchronously
only once the session bus is ready), and in such case we should just
ignore any request involving the display configuration, given that it will
be performed again once we've it ready. Otherwise we could crash.

LP: #1887731