Remove pop-shell.patch
This is unnecessary if pop-shell installs it's schemas in the ordinary
system directory, and provides the `.xml` files for gnome-control-center
to show these key bindings.

This requires a corresponding change in the pop-shell package. But it is
not strictly breaking; it will just means the pop-shell bindings will
not be shown.