Revert the "Fix locales updates" commit
Commit 2ba4e350 changed the behavior when changing the display language
from Settings in an undesired way.

In Ubuntu we have seen Language and Format as two separate things, i.e.
when changing the Language, the effective Format has not been changed
and vice versa. Ubuntu specific patches in accountsservice make it work
that way.

But commit 2ba4e350 enforces the locale selected for the new Language
to be applied also for Format. That behavior is unintuitive, since the
GUI does not indicate such a relation between those two settings. It
also differs from how it works if you use Language Support to change
the display language.

This reversal changes it back to the previous behavior.