multitasking-panel: Sync workspace and monitor isolation dock settings
When using ubuntu-dock the multitasking panel setting doesn't fully
apply, because even though the user selects to isolate workspaces, the
dock may still continue to show icons from all the workspaces.

As per this, ensure that the settings are applied also to the dock, plus
include dock-specifc settings to control multi-monitor behavior.