ubuntu-panel: Do not change themes on startup (or on user changes)
Do not override user themes unless an user has explicitly selected a
color theme. This applies to both the user theme and the icons, as we
want a consistent experience. But users could still use tweaks to change
them.

At the same time, avoid changing the gedit color scheme if an user
previously set a non-Yaru theme.

LP: #1968213